Clean up ResultSetFormatter
It now operates in two phases. The first phase is taking the result set and converting it into a number of human-readable fragments. Depending on whether the user is a console or a player, those fragments are either transmitted as colorized text or JSON. The way the fragments are rearranged are separate classes, cleaning up the implementation of ResultSetFormatter and associated classes signficantly.